create score file

+++ test/mmw/mmw-score.bats
@@ -17,7 +17,10 @@ teardown() {
   mmw-config addProjFilePathDeclarationAndDefinition
   mmw-config addArtistDeclaration
   mmw setArtist test-artist
-  rm -r $HOME/audio/test-artist/My_Test_MMW_project/
+  if [ -d $HOME/audio/test-artist/My_Test_MMW_project/ ]; then 
+    rm -r $HOME/audio/test-artist/My_Test_MMW_project/
+  fi
+
   echo "y" | mmw createProj
   [ -d "$(mmw displayProj)" ] # PROJPATH
   [ -f $HOME/audio/test-artist/My_Test_MMW_project/stems/input_1.wav ]            
@@ -26,9 +29,22 @@ teardown() {
 }
 
 @test "createScoreDesign should create the project score file" {
-  source $HOME/.config/mmw.config
-  mmw createLyScoreAndMidiFile
-  [ -f "$LILY" ]
+  mmw-config addProjNameDeclaration
+  mmw setProj My_Test_MMW_project
+  mmw-config addProjLocationDeclaration 
+  mmw setProjLoc mmw
+  mmw-config addProjFilePathDeclarationAndDefinition
+  mmw-config addArtistDeclaration
+  mmw setArtist test-artist
+  mmw-config addScoreDeclarationAndDefinition
+  if [ -d $HOME/audio/test-artist/My_Test_MMW_project/ ]; then 
+    rm -r $HOME/audio/test-artist/My_Test_MMW_project/
+  fi
+  echo "y" | mmw createProj
+  [ "$HOME/mmw/My_Test_MMW_project.ly" = "$(mmw getScore)" ]
+  mmw writeLilyFile
+  [ -f "$(mmw getScore)" ]
+
   #[ -f "$MIDI" ]
   #[ -f "$PDF" ]
